--- src/components/popups/PopUser/PopUser.jsx
@@ -1,9 +1,0 @@
-import "./PopUser.css";
-
-function PopUser() {
-  // Этот компонент больше не используется, так как у нас есть отдельная страница ExitPage
-  // Можно оставить пустым или удалить совсем
-  return null;
-}
-
-export default PopUser;
