+++ src/utils/formatMoney.js
@@ -47,25 +47,6 @@ export const formatMoney = (value) => {
 };
 
 /**
- * Форматирует сумму для отображения с разделителями тысяч и символом рубля
- * @param {string|number} value - Значение для форматирования
- * @returns {string} Отформатированная сумма с символом рубля
- */
-export const formatMoneyDisplay = (value) => {
-  if (!value) return "";
-
-  const numericValue = parseFloat(value);
-  if (isNaN(numericValue)) return "";
-
-  return new Intl.NumberFormat("ru-RU", {
-    style: "currency",
-    currency: "RUB",
-    minimumFractionDigits: 2,
-    maximumFractionDigits: 2,
-  }).format(numericValue);
-};
-
-/**
  * Валидирует введенное значение суммы
  * @param {string} value - Значение для валидации
  * @returns {boolean} true если значение корректно
